Add procfile
following the steps outlined here:
https://devcenter.heroku.com/articles/django-app-configuration
A procfile tells Heroku what to "run" when you push your project
to their servers.

"gunicorn" is a program that runs "servers" in production. it is roughly
equivalent to "python manage.py runserver" which you have been using in
development.

The next bit - in the tutorial is "myproject.wsgi". This for us is
"test.wsgi" b/c in this project I named the app "test". That was
probably a bad choice of names, b/c it will get confusing later

wsgi is a file that tells gunicorn how to setup the app, and the name
"test.wsgi" is saying look in the "test" module (a module is a folder
with __init__.py in it, and look for the "wsgi" file.